refactor: Make DDL preprocessing generic for every DataFusion catalog
