+++ crates/runtime/Cargo.toml
@@ -41,7 +41,7 @@ chunking = { path = "../chunking" }
 clap.workspace = true
 csv = "1.3.0"
 dashmap = "6.1.0"
-data_components = { path = "../data_components" }
+data_components = { path = "../data_components", features = ["flightsql"] }
 dataformat-json = { path = "../dataformat-json" }
 datafusion.workspace = true
 datafusion-datasource.workspace = true
@@ -238,7 +238,6 @@ repl = { path = "../repl" }
 rstest = "0.26.1"
 scopeguard = "1.2.0"
 serde_json = { workspace = true, features = ["preserve_order"] }
-yaml = { path = "../yaml" }
 spice-cloud = { path = "../spice_cloud" }
 spiceai.workspace = true
 tar = "0.4"
@@ -248,6 +247,7 @@ tiberius.workspace = true
 tokio = { workspace = true, features = ["time", "test-util"] }
 tracing-opentelemetry.workspace = true
 tracing-subscriber.workspace = true
+yaml = { path = "../yaml" }
 
 [features]
 aws-secrets-manager = ["dep:aws-sdk-sts", "runtime-secrets/aws-secrets-manager"]
@@ -325,7 +325,11 @@ sftp = ["runtime-object-store/sftp"]
 sharepoint = ["data_components/sharepoint"]
 smb = ["runtime-object-store/smb"]
 snapshots = ["runtime-acceleration/snapshots"]
-snowflake = ["dep:snowflake-api", "db_connection_pool/snowflake", "data_components/snowflake"]
+snowflake = [
+    "dep:snowflake-api",
+    "db_connection_pool/snowflake",
+    "data_components/snowflake",
+]
 spark = ["data_components/spark_connect"]
 sqlite = [
     "dep:tokio-rusqlite",
