+++ crates/runtime/src/datafusion/builder.rs
@@ -422,7 +422,7 @@ impl DataFusionBuilder {
         let caching = self.caching.unwrap_or(Arc::new(Caching::default()));
 
         let ddl_enabled_catalogs = Arc::new(RwLock::new(HashSet::new()));
-        let ddl_options_store = super::iceberg_ddl::acceleration_options::new_shared_store();
+        let ddl_extension_store = super::ddl::acceleration_options::new_shared_store();
 
         // Add the Iceberg DDL analyzer rule after context creation so it can
         // reference the catalog list and DDL-enabled catalogs.
@@ -432,7 +432,7 @@ impl DataFusionBuilder {
             super::iceberg_ddl::analyzer_rule::IcebergDdlAnalyzerRule::new(
                 ctx.state().catalog_list(),
                 &ddl_enabled_catalogs,
-                Arc::clone(&ddl_options_store),
+                Arc::clone(&ddl_extension_store),
             ),
         ));
 
@@ -451,7 +451,7 @@ impl DataFusionBuilder {
             data_writers: RwLock::new(HashSet::new()),
             writable_catalogs: RwLock::new(HashSet::new()),
             ddl_enabled_catalogs,
-            ddl_options_store,
+            ddl_extension_store,
             datafusion_ref,
             caching,
             pending_sink_tables: TokioRwLock::new(Vec::new()),
