+++ crates/runtime/src/datafusion/query.rs
@@ -936,11 +936,11 @@ impl Query {
         let plan = match self.sql {
             QueryMethod::Plan(ref plan) => plan.clone(),
             QueryMethod::Text { ref sql, .. } => {
-                // Pre-process CREATE TABLE ... WITH (acceleration.*, dataset.*) before planning
+                // Pre-process CREATE TABLE extensions (WITH options, PARTITION BY) before planning
                 let preprocessed =
-                    match super::iceberg_ddl::preprocess::preprocess_create_table_with_options(
+                    match super::ddl::preprocess::preprocess_create_table_with_options(
                         sql,
-                        self.df.ddl_options_store(),
+                        self.df.ddl_extension_store(),
                     ) {
                         Ok(preprocessed) => preprocessed,
                         Err(e) => {
@@ -951,11 +951,11 @@ impl Query {
                     };
 
                 let (effective_sql, store_key) = match &preprocessed {
-                    super::iceberg_ddl::preprocess::PreprocessResult::Modified {
+                    super::ddl::preprocess::PreprocessResult::Modified {
                         sql: modified,
                         store_key,
                     } => (modified.as_str(), Some(store_key.as_str())),
-                    super::iceberg_ddl::preprocess::PreprocessResult::Unchanged => {
+                    super::ddl::preprocess::PreprocessResult::Unchanged => {
                         (sql.as_ref(), None)
                     }
                 };
@@ -965,8 +965,8 @@ impl Query {
                     Err(e) => {
                         if let Some(store_key) = store_key
                             && let Err(cleanup_err) =
-                                super::iceberg_ddl::preprocess::cleanup_preprocessed_ddl_options(
-                                    self.df.ddl_options_store(),
+                                super::ddl::preprocess::cleanup_preprocessed_ddl_options(
+                                    self.df.ddl_extension_store(),
                                     store_key,
                                 )
                         {
